Final changes, everything works
