+++ src/pages/generator.tsx
@@ -1,5 +1,6 @@
 import dynamic from 'next/dynamic'
 import styled from 'styled-components'
+import { useEffect } from 'react'
 import { FormProvider, useForm } from 'react-hook-form'
 import { FormValues } from '../types'
 import { Form } from '../components/generator/form/Form'
@@ -35,6 +36,19 @@ const initialFormValues = {
 
 export default function GeneratorPage() {
   const formContext = useForm<FormValues>({ defaultValues: initialFormValues })
+  // TODO: move this to a custom react hook
+  useEffect(() => {
+    const lastSession = localStorage.getItem('jsonResume')
+    if (lastSession) {
+      // TODO: validate JSON schema using Zod
+      const jsonResume = JSON.parse(lastSession) as FormValues
+      formContext.reset(jsonResume)
+    }
+    const subscription = formContext.watch((data) => {
+      localStorage.setItem('jsonResume', JSON.stringify(data))
+    })
+    return () => subscription.unsubscribe()
+  }, [formContext])
   return (
     <Main>
       <FormProvider {...formContext}>
